requirements.txt
ruff==0.4.3
future==1.0.0
mypy==1.4.1
ableton-control-surface-core==0.0.5